- Add SearchRepository Class - Add LegoSet data class - Add LegoTheme data class - Use find function instead of for loops
