+++ src/main/kotlin/LegoSet.kt
@@ -1,0 +1,7 @@
+data class LegoSet(
+    val set_id : String,
+    val name: String,
+    val year: String,
+    val theme_id: String,
+    val num_parts: String
+)
