+++ src/main/kotlin/LegoTheme.kt
@@ -1,0 +1,5 @@
+data class LegoTheme(
+    val id: String,
+    val name: String,
+    val parent_id: String?
+)
